[CheckPoint] destroy code which began to move rails engine.

+++ db/principals_migrate/20260415000002_remove_redundant_index_from_user_app_preferences.rb
@@ -2,6 +2,9 @@
 
 class RemoveRedundantIndexFromUserAppPreferences < ActiveRecord::Migration[8.2]
   def change
+    return unless table_exists?(:user_app_preferences)
+    return if !column_exists?(:user_app_preferences, :user_id) || !index_exists?(:user_app_preferences, :user_id)
+
     remove_index(:user_app_preferences, :user_id)
   end
 end
